Split Key file from Pitch file
